app2/src/data_service/marketplace/sql_generator/requirements.txt
boto3~=1.28.67
jinja2==3.1.2
jsonref~=1.1.0
pydantic~=2.4.2
requests
pyyaml==6.0.1
snowflake-connector-python
awslambdaric
aws-cdk-lib
dataclasses-json==0.6.1
nodeenv~=1.8.0